data-layer/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
numpy
tensorflow
tf-keras
transformers
python-dotenv
python-multipart
langchain
openai
gpt4all
langchain-community
sentence-transformers
langchain_postgres
PyPDF2
python-docx
langchain-ollama
langchain-google-genai
pillow
starlette